perf(bundle): more lazy-loads + extract trigger editor
Convert remaining customElement re-exports in ui/index.ts to type-only
(AgentInterface, MessageEditor, MessageList, StreamingMessageContainer,
ThinkingBlock, DiffBlock, ErrorMessage, ErrorDetails, ExpandableSection,
SkillChip, Input, ConsoleBlock, CustomProviderCard, BgProcessPill,
AttachmentTile, ContinueSessionChooser, PersistentStorageDialog,
SessionListDialog, VerificationOutputModal, ProjectPickerPopover,
SandboxedIframe, SearchBox, SearchResults, message classes). The actual
customElements.define(...) registrations still happen via static
side-effect imports inside the components themselves \u2014 type-only
keeps the public library type surface without dragging the value graph
into entry chunks that only need an unrelated symbol like AppStorage.

Lazy-load five more conditional-render LitElements via
`app/lazy-widgets.ts`:
- BgProcessPill (only when bg processes exist)
- CostPopover (only on cost button click)
- PreCompactionHistory (only when compaction occurs)
- SearchBox + SearchResults (mobile landing / sidebar search)
- ContinueSessionChooser (only on continue-session prompt)
- ProjectPickerPopover (only on +New Goal with \u22652 projects)

Extract the staff-assistant trigger editor (\~200 LOC, 10 kB of
templates) from render.ts into render-triggers.ts; render.ts now
lazy-loads the module the first time a staff proposal panel renders
`<triggers-editor>`. Pure-function fallbacks keep `parseTriggers`
and `hasInvalidGoalTriggersForPreview` working before the chunk lands.

Entry chunk: 703 kB \u2192 618 kB raw, 184 kB \u2192 162 kB gz (cumulative
across this commit + prior).
