+++ src/app/lazy-widgets.ts
@@ -34,3 +34,47 @@ export function loadAttachmentOverlay(): Promise<typeof import("../ui/dialogs/At
 	_attachmentOverlay = import("../ui/dialogs/AttachmentOverlay.js");
 	return _attachmentOverlay;
 }
+
+// ── Conditional-render LitElements ─────────────────────────────────
+// These tags only appear after a user action / runtime event — bg
+// processes spawn, a popover opens, search bar mounts, compaction
+// completes, the agent prompts to continue a session, etc.
+// Fire-and-forget triggers; Lit upgrades the unknown tag once defined.
+
+let _bgProcessPill: Promise<unknown> | null = null;
+export function ensureBgProcessPill(): Promise<unknown> {
+	if (_bgProcessPill) return _bgProcessPill;
+	_bgProcessPill = import("../ui/components/BgProcessPill.js");
+	return _bgProcessPill;
+}
+
+let _costPopover: Promise<unknown> | null = null;
+export function ensureCostPopover(): Promise<unknown> {
+	if (_costPopover) return _costPopover;
+	_costPopover = import("../ui/components/CostPopover.js");
+	return _costPopover;
+}
+
+let _preCompactionHistory: Promise<unknown> | null = null;
+export function ensurePreCompactionHistory(): Promise<unknown> {
+	if (_preCompactionHistory) return _preCompactionHistory;
+	_preCompactionHistory = import("../ui/components/PreCompactionHistory.js");
+	return _preCompactionHistory;
+}
+
+let _searchBox: Promise<unknown> | null = null;
+export function ensureSearchBox(): Promise<unknown> {
+	if (_searchBox) return _searchBox;
+	_searchBox = Promise.all([
+		import("../ui/components/SearchBox.js"),
+		import("../ui/components/SearchResults.js"),
+	]);
+	return _searchBox;
+}
+
+let _continueSessionChooser: Promise<unknown> | null = null;
+export function ensureContinueSessionChooser(): Promise<unknown> {
+	if (_continueSessionChooser) return _continueSessionChooser;
+	_continueSessionChooser = import("../ui/components/ContinueSessionChooser.js");
+	return _continueSessionChooser;
+}
